requirements.txt
python-telegram-bot[ext]
gspread
google-auth-oauthlib
google-auth-httplib2
thefuzz
requests
pandas
gspread-dataframe